[refactor] 헷갈리지 않게 shell_test.py 이름 변경
